Refactors models out of __init__.py
