+++ app/__init__.py
@@ -2,19 +2,13 @@ import os
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from app.models import db
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False #Handles deprication warning
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ["DATABASE_URL"]
-db = SQLAlchemy(app)
+db.init_app(app)
 migrate = Migrate(app, db)
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(128))
-    
-    def __repr__(self):
-        return '<User %r>' % self.name
 
 
 @app.route('/')
