+++ app/models.py
@@ -1,0 +1,11 @@
+from flask_sqlalchemy import SQLalchemy
+
+db = SQLAlchemy()
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(128))
+
+    def __repr__(self):
+        return '<User %r>' % self.name
+
